+++ Addons/Schumix.CompilerAddon/CompilerAddon.cs
@@ -44,6 +44,7 @@ namespace Schumix.CompilerAddon
 	{
 		private readonly LocalizationConsole sLConsole = Singleton<LocalizationConsole>.Instance;
 		private readonly LocalizationManager sLManager = Singleton<LocalizationManager>.Instance;
+		private readonly Runtime sRuntime = Singleton<Runtime>.Instance;
 		private readonly IrcBase sIrcBase = Singleton<IrcBase>.Instance;
 		private readonly Regex regex = new Regex(@"^\{(?<code>.*)\}$");
 		private SCompiler sSCompiler;
@@ -168,7 +169,7 @@ namespace Schumix.CompilerAddon
 
 			if(CompilerConfig.MaxAllocatingE)
 			{
-				var memory = Process.GetCurrentProcess().WorkingSet64/1024/1024;
+				var memory = sRuntime.MemorySizeToMB;
 				int ircnetwork = sIrcBase.Networks.Count > 1 ? 20 * sIrcBase.Networks.Count : 0;
 
 				if(memory > CompilerConfig.MaxAllocatingM + ircnetwork)